backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
elasticsearch==8.12.0
python-dateutil==2.8.2
kafka-python==2.0.2
